Adds some reprotest scripts
Curious just how reproducible cargo packages are by default. Looks like
they're path-dependent, which is a shame, but remapping paths for the
release binaries can be done via an override.

Unfortunately, this override must be dynamic in terms of the current
env: it must reference $HOME and $CARGO_HOME, for example. Exporting
remap-path-prefix as RUSTFLAGS via env var works, but requires an
out-of-build-toolchain action, i.e. sourcing the vars, for it to take
effect.