Applies nightly lints
Ran the nightly linter, which pointed out a bunch of bad patterns that
the stable linter didn't. Followed suit and removed a bit of "clone"
overuse, as I get more comfortable with using references judiciously.
Still convinced by interaction of static lifetimes and sync, so leaving
some clones in place for now.