Ensure Wireguard subnets are /30
The introduction of ipgen meant that IPv4 address pairs
for a given service were not guaranteed to be adjacent, which means they
aren't a /30, which broke networking. While deterministic IP generation
is a great property, it's more important for the tool to figure out
what's unused and use that.

Move error to its own file, seems idiomatic.