Post ephemeral SSH key to DO account
The DO server always had an ephemeral SSH keypair created just for it,
configured via cloudinit, but DigitalOcean sends an email with a root
password by default on new instance creation, which led to a lot of
unwanted messages. If we also supply the keypair as an API object to DO,
then no emails are sent on instance creation. Hurray.

Took the opportunity to reorg the server logic into its own module.
Still learning best practices for nested modules in rust. Ideally
there'd be another layer of abstraction for "providers", e.g. AWS,
but no need to do that now, since only one is supported.